Backfill tests for Biloba reporter
The tests integrate with both the ginkgo CLI and the test2json tool.
I've taken steps to account for variation between runs and between
computers (i.e. replacing duration strings with "TIME" and fetching the
repo root dynamically). Even with this effort I still expect that this
test may fail in other environments (e.g. on linux). We'll have to
address those problems as they occur

Authored-by: Matt Royal <mroyal@pivotal.io>